Created a base class for events caused by a user
